+++ hs-bindgen/src-internal/HsBindgen/IR/C/Type.hs
@@ -40,6 +40,7 @@ module HsBindgen.IR.C.Type (
 
     -- * Normal forms
   , Normalize(..)
+  , getCanonicalType
 
     -- * Queries
   , ValOrRef(..)