+++ hs-bindgen/test-artefacts/headers/golden/binding-specs/rep/emptydata/opaque.h
@@ -1,5 +1,0 @@
-struct foo;
-
-struct bar {
-  struct foo *a;
-};