+++ hs-bindgen/test-artefacts/headers/golden/binding-specs/rep/emptydata/typedef.h
@@ -1,3 +1,0 @@
-#include "pt.h"
-
-typedef struct pt origin;